+++ app/src/main/java/chat/rocket/android/contacts/ContactsFragment.kt
@@ -1,14 +1,11 @@
 package chat.rocket.android.contacts
 
 import android.Manifest
-import android.content.Intent
 import android.content.pm.PackageManager
 import android.graphics.Color
-import android.net.Uri
 import android.os.Bundle
 import android.view.*
 import android.widget.ImageView
-import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.SearchView
 import androidx.core.content.ContextCompat
@@ -17,15 +14,12 @@ import chat.rocket.android.R
 import chat.rocket.android.contacts.models.Contact
 import chat.rocket.android.main.ui.MainActivity
 import chat.rocket.android.util.extension.onQueryTextListener
-import dagger.android.support.AndroidSupportInjection
 import kotlinx.android.synthetic.main.app_bar.*
 import java.util.ArrayList
 import kotlin.collections.HashMap
 
 // WIDECHAT
 import chat.rocket.android.helper.Constants
-import chat.rocket.android.server.domain.GetAccountInteractor
-import chat.rocket.android.server.domain.GetCurrentServerInteractor
 import com.facebook.drawee.view.SimpleDraweeView
 import android.view.LayoutInflater
 import android.widget.TextView
@@ -45,48 +39,28 @@ import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.rxkotlin.subscribeBy
 import io.reactivex.schedulers.Schedulers
 import javax.inject.Inject
-import com.google.firebase.dynamiclinks.DynamicLink
-import com.google.firebase.dynamiclinks.FirebaseDynamicLinks
-import com.google.firebase.dynamiclinks.ShortDynamicLink
-import kotlinx.android.synthetic.main.fragment_contact_parent.view.*
-import kotlinx.coroutines.experimental.launch
-import javax.inject.Inject
 
 /**
  * Load a list of contacts in a recycler view
  */
 class ContactsFragment : Fragment() {
 
-	@Inject
-	lateinit var serverInteractor: GetCurrentServerInteractor
-
-	@Inject
-	lateinit var getAccountInteractor: GetAccountInteractor
-
-	/**
     @Inject
     lateinit var dbFactory: DatabaseManagerFactory
+
     @Inject
     lateinit var serverInteractor: GetCurrentServerInteractor
+
     private var recyclerView :RecyclerView? = null
     private var emptyTextView:  TextView? = null
-
-    private val MY_PERMISSIONS_REQUEST_RW_CONTACTS = 0
 
     /**
      * The list of contacts to load in the recycler view
      */
     private var contactArrayList: ArrayList<Contact> = ArrayList()
 
-    /**
-     *  The mapping of contacts with their registration status
-     */
-    private var contactHashMap: HashMap<String, String> = HashMap()
-
     private val MY_PERMISSIONS_REQUEST_RW_CONTACTS = 0
 
-    private var createNewChannelLink: View? = null
-    private var shareViaAnotherApp: View? = null
     private var searchView: SearchView? = null
     private var sortView: MenuItem? = null
     private var searchIcon: ImageView? = null
@@ -139,37 +113,6 @@ class ContactsFragment : Fragment() {
         this.recyclerView = view.findViewById(R.id.recycler_view)
         this.emptyTextView = view.findViewById(R.id.text_no_data_to_display)
         getContactsPermissions()
-        
-        shareViaAnotherApp = view.findViewById(R.id.share_via_another_app)
-		shareViaAnotherApp!!.setOnClickListener { _ ->
-
-			FirebaseDynamicLinks.getInstance().createDynamicLink()
-				.setLink(Uri.parse("https://open.rocket.chat/direct/Shailesh351"))
-				.setDomainUriPrefix("https://dylinktesting.page.link")
-				.setAndroidParameters(
-					DynamicLink.AndroidParameters.Builder("chat.veranda.android.dev").build())
-				.setSocialMetaTagParameters(
-					DynamicLink.SocialMetaTagParameters.Builder()
-						.setTitle("Shailesh351")
-						.setDescription("Chat with Shailesh on Rocket.Chat")
-						.build())
-				.buildShortDynamicLink(ShortDynamicLink.Suffix.SHORT)
-				.addOnSuccessListener { result ->
-					val link = result.shortLink.toString()
-					Toast.makeText(context, link, Toast.LENGTH_SHORT).show()
-
-					val shareIntent = Intent()
-					shareIntent.action = Intent.ACTION_SEND
-					shareIntent.putExtra(Intent.EXTRA_TEXT, "Default Invitation Text : $link")
-					shareIntent.type = "text/plain"
-					startActivity(shareIntent, null)
-
-				}.addOnFailureListener {
-					// Error
-					Toast.makeText(context, "Error dynamic link", Toast.LENGTH_SHORT).show()
-				}
-		}
-        
         return view
     }
 
@@ -401,82 +344,4 @@ class ContactsFragment : Fragment() {
         finalList.add(inviteItemHolder("invite"))
         return finalList
     }
-
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        val view = inflater.inflate(R.layout.fragment_contact_parent, container, false)
-
-        createNewChannelLink = view.findViewById(R.id.create_new_channel_button)
-        createNewChannelLink!!.setOnClickListener {
-            val createChannelFragment = CreateChannelFragment()
-            val transaction = activity?.supportFragmentManager?.beginTransaction();
-            transaction?.setCustomAnimations(R.anim.enter_from_right, R.anim.exit_to_left, R.anim.enter_from_left, R.anim.exit_to_right);
-            transaction?.replace(this.id, createChannelFragment, "createChannelFragment");
-            transaction?.addToBackStack(null)?.commit();
-        }
-
-		shareViaAnotherApp = view.findViewById(R.id.share_via_another_app)
-		shareViaAnotherApp!!.setOnClickListener { _ ->
-
-			launch {
-				//get serverUrl and username
-				val server = serverInteractor.get()!!
-				val account = getAccountInteractor.get(server)!!
-				val userName = account.userName
-
-				FirebaseDynamicLinks.getInstance().createDynamicLink()
-					.setLink(Uri.parse("$server/direct/$userName"))
-					.setDomainUriPrefix("https://dylinktesting.page.link")
-					.setAndroidParameters(
-						DynamicLink.AndroidParameters.Builder("chat.veranda.android.dev").build())
-					.setSocialMetaTagParameters(
-						DynamicLink.SocialMetaTagParameters.Builder()
-							.setTitle(userName)
-							.setDescription("Chat with $userName on Rocket.Chat")
-							.build())
-					.buildShortDynamicLink(ShortDynamicLink.Suffix.SHORT)
-					.addOnSuccessListener { result ->
-						val link = result.shortLink.toString()
-						Toast.makeText(context, link, Toast.LENGTH_SHORT).show()
-
-						val shareIntent = Intent()
-						shareIntent.action = Intent.ACTION_SEND
-						shareIntent.putExtra(Intent.EXTRA_TEXT, "Default Invitation Text : $link")
-						shareIntent.type = "text/plain"
-						startActivity(shareIntent, null)
-
-					}.addOnFailureListener {
-						// Error
-						Toast.makeText(context, "Error dynamic link", Toast.LENGTH_SHORT).show()
-					}
-			}
-		}
-
-		return view
-	}
-
-    companion object {
-
-        /**
-         * Create a new ContactList fragment that displays the given list of contacts
-         *
-         * @param contactArrayList the list of contacts to load in the recycler view
-         * @param contactHashMap the mapping of contacts with their registration status
-         * @return the newly created ContactList fragment
-         */
-        fun newInstance(
-                contactArrayList: ArrayList<Contact>,
-                contactHashMap: HashMap<String, String>
-        ): ContactsFragment {
-            val contactsFragment = ContactsFragment()
-
-            val arguments = Bundle()
-            arguments.putParcelableArrayList("CONTACT_ARRAY_LIST", contactArrayList)
-            arguments.putSerializable("CONTACT_HASH_MAP", contactHashMap)
-
-            contactsFragment.arguments = arguments
-
-            return contactsFragment
-        }
-    }
-
 }
