--- app/src/main/java/chat/rocket/android/contacts/di/ContactsFragmentProvider.kt
@@ -1,12 +1,0 @@
-package chat.rocket.android.contacts.di
-
-import chat.rocket.android.contacts.ContactsFragment
-import dagger.Module
-import dagger.android.ContributesAndroidInjector
-
-@Module
-abstract class ContactsFragmentProvider {
-
-	@ContributesAndroidInjector()
-	abstract fun provideContactsFragment(): ContactsFragment
-}
