+++ app/src/main/res/layout/fragment_contact_parent.xml
@@ -6,22 +6,20 @@
     android:layout_height="match_parent"
     tools:context=".contacts.ContactsFragment">
 
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/recycler_view"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            tools:listitem="@layout/item_contact" />
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/recycler_view"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        tools:listitem="@layout/item_contact" />
 
-        <TextView
-            android:id="@+id/text_no_data_to_display"
-            style="@style/TextAppearance.AppCompat.Subhead"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:gravity="center"
-            android:text="@string/msg_no_data_to_display"
-            android:textSize="24dp"
-            android:visibility="gone"
-            tools:visibility="visible" />
+    <TextView
+        android:id="@+id/text_no_data_to_display"
+        style="@style/TextAppearance.AppCompat.Subhead"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/msg_no_data_to_display"
+        android:visibility="gone"
+        tools:visibility="visible" />
     <View
         android:id="@+id/divider3"
         android:layout_width="match_parent"
@@ -30,36 +28,5 @@
         android:layout_marginTop="16dp"
         android:visibility="gone"
         android:background="?android:attr/listDivider" />
-
-    <View
-        android:id="@+id/divider"
-        android:layout_width="match_parent"
-        android:layout_height="1dp"
-        android:layout_above="@id/share_via_another_app"
-        android:background="?android:attr/listDivider" />
-
-    <LinearLayout
-        android:id="@+id/share_via_another_app"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:layout_marginTop="12dp"
-        android:layout_marginBottom="12dp">
-
-        <ImageView
-            android:layout_width="40dp"
-            android:layout_height="40dp"
-            android:layout_marginStart="10dp"
-            android:contentDescription="@string/share_via_another_app"
-            android:src="@drawable/ic_invite_black_24dp" />
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:layout_marginStart="12dp"
-            android:text="@string/share_via_another_app"
-            android:textSize="18sp" />
-    </LinearLayout>
 
 </RelativeLayout>
